# openapi.yaml
openapi: 3.0.1
info:
  title: Trabajo Individual - Juan Molina
  version: v0.0.1
paths:
  /api/v1/sign_up:
    post:
      tags:
        - Controller
      operationId: signUpUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserEntity'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorsDTO'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorsDTO'
  /api/v1/login:
    post:
      tags:
        - Controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserEntity'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorsDTO'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorsDTO'
  /api/v1/refreshToken:
    get:
      tags:
        - Controller
      operationId: refreshToken
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorsDTO'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorsDTO'
components:
  schemas:
    ErrorsDTO:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ExceptionDTO'
    ExceptionDTO:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        code:
          type: integer
          format: int32
        details:
          type: string
    PhoneEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        cityCode:
          type: integer
          format: int32
        countryCode:
          type: string
    UserDTO:
      required:
        - email
        - password
      type: object
      properties:
        name:
          type: string
        email:
          pattern: .+@.+\..+
          type: string
        password:
          maxLength: 12
          minLength: 8
          type: string
        phones:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/PhoneEntity'
    UserEntity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phones:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/PhoneEntity'
        token:
          type: string
        active:
          type: boolean